refactor views

+++ views/game/deleteGame.py
@@ -1,6 +1,7 @@
 from main import app, db
-from flask import flash, redirect, url_for
-from models import Games
+from flask import flash, redirect, url_for, session
+from models import Games, Users
+from werkzeug.exceptions import NotFound
 import os
 from helpers import delete_image
 
@@ -8,17 +9,26 @@ from helpers import delete_image
 """View para deletar jogo do Banco de dados"""
 @app.route('/deletar/<int:id_game>')
 def delete_game(id_game):
-    game = Games.query.filter_by(id=id_game)
-    url_game = game.first().url_image
-    game.delete()
+    user_nickname = session['user_logged']
+
     try:
-        delete_image(url_game)
+        user = Users.query.filter_by(nickname=user_nickname).first_or_404()
+        if user.is_adm:
+            game = Games.query.filter_by(id=id_game)
+            url_game = game.first_or_404().url_image
+            game.delete()
+            delete_image(url_game)
+            db.session.add(user)
+
+    except NotFound:
+        flash('Exclusão de Jogo não realizda','danger')
+        return redirect(url_for('index'))
+
     except FileNotFoundError as err:
         print(err, 'Jogo sem imagem cadastrada')
 
     except TypeError as err:
         print(err, 'Jogo sem imagem')
-
     else:
         print('Exclusão da imagem do jogo realizada com sucesso')
